refactor: complete split of api/server.ts into 12 route modules
Split 3032-line monolith into focused route files:
- server.ts now 349 lines (thin shell: express setup, auth, route registration)
- 12 route files in src/api/routes/ totaling 3237 lines
- context.ts provides RouteContext interface + shared helpers

Route files: browser (14), tabs (7), snapshots (8), devtools (15),
extensions (14), network (10), sessions (11), agents (15), data (25),
content (14), media (19), misc (58) — 210 routes total.

All tests passing. Implements Item 1 of docs/STRUCTURE-IMPROVEMENTS.md.
